backend/requirements.txt
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djoser==2.2.0
psycopg2==2.9.7
python-dotenv==1.0.0
Pillow==10.0.0
flake8==6.0.0
flake8-isort==6.0.0
isort==5.12.0
requests==2.26.0
gunicorn==20.1.0